add fuctions to FrameResourceManager class

+++ Source/Public/App/Settings.h
@@ -4,7 +4,7 @@
 
 struct Settings
 {
-    static constexpr int mNumFrameResources = 3;
+    static constexpr size_t mNumFrameResources = 3;
     static inline UINT mWidth = 1600;
     static inline UINT mHeight = 800;
